requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
psycopg2-binary
sendgrid
aiohttp
schedule
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
alembic